requirements.txt
pytest==7.1.2
selenium==4.4.0
allure-pytest==2.11.0